Move Stackage class definition to a header file.
