+++ include/rospack/rospack.h
@@ -111,6 +111,7 @@ and Rosstack.
 #include <map>
 #include <set>
 #include <string>
+#include "tinyxml.h"
 #include <vector>
 #include "macros.h"
 
@@ -130,8 +131,39 @@ typedef enum
 } traversal_order_t;
 
 // Forward declarations
-class Stackage;
 class DirectoryCrawlRecord;
+
+class Stackage
+{
+  public:
+    // \brief name of the stackage
+    std::string name_;
+    // \brief absolute path to the stackage
+    std::string path_;
+    // \brief absolute path to the stackage manifest
+    std::string manifest_path_;
+    // \brief filename of the stackage manifest
+    std::string manifest_name_;
+    // \brief package's license with a support for multi-license.
+    std::vector<std::string> licenses_;
+    // \brief have we already loaded the manifest?
+    bool manifest_loaded_;
+    // \brief TinyXML structure, filled in during parsing
+    TiXmlDocument manifest_;
+    std::vector<Stackage*> deps_;
+    bool deps_computed_;
+    bool is_wet_package_;
+    bool is_metapackage_;
+
+    Stackage(const std::string& name,
+             const std::string& path,
+             const std::string& manifest_path,
+             const std::string& manifest_name);
+
+    void update_wet_information();
+    bool isStack() const;
+    bool isPackage() const;
+};
 
 /**
  * @brief The base class for package/stack ("stackage") crawlers.  Users of the library should
