+++ src/rospack.cpp
@@ -136,29 +136,7 @@ class Exception : public std::runtime_error
     {}
 };
 
-class Stackage
-{
-  public:
-    // \brief name of the stackage
-    std::string name_;
-    // \brief absolute path to the stackage
-    std::string path_;
-    // \brief absolute path to the stackage manifest
-    std::string manifest_path_;
-    // \brief filename of the stackage manifest
-    std::string manifest_name_;
-    // \brief package's license with a support for multi-license.
-    std::vector<std::string> licenses_;
-    // \brief have we already loaded the manifest?
-    bool manifest_loaded_;
-    // \brief TinyXML structure, filled in during parsing
-    TiXmlDocument manifest_;
-    std::vector<Stackage*> deps_;
-    bool deps_computed_;
-    bool is_wet_package_;
-    bool is_metapackage_;
-
-    Stackage(const std::string& name,
+Stackage::Stackage(const std::string& name,
              const std::string& path,
              const std::string& manifest_path,
              const std::string& manifest_name) :
@@ -173,45 +151,43 @@ class Stackage
       is_wet_package_ = manifest_name_ == ROSPACKAGE_MANIFEST_NAME;
     }
 
-    void update_wet_information()
+void Stackage::update_wet_information()
+{
+  assert(is_wet_package_);
+  assert(manifest_loaded_);
+  // get name from package.xml instead of folder name
+  TiXmlElement* root = get_manifest_root(this);
+  for(TiXmlElement* el = root->FirstChildElement("name"); el; el = el->NextSiblingElement("name"))
+  {
+    name_ = el->GetText();
+    break;
+  }
+  // Get license texts, where there may be multiple elements for.
+  std::string tagname_license = "license";
+  for(TiXmlElement* el = root->FirstChildElement(tagname_license); el; el = el->NextSiblingElement(tagname_license ))
+  {
+    licenses_.push_back(el->GetText());
+  }
+  // check if package is a metapackage
+  for(TiXmlElement* el = root->FirstChildElement("export"); el; el = el->NextSiblingElement("export"))
+  {
+    if(el->FirstChildElement("metapackage"))
     {
-      assert(is_wet_package_);
-      assert(manifest_loaded_);
-      // get name from package.xml instead of folder name
-      TiXmlElement* root = get_manifest_root(this);
-      for(TiXmlElement* el = root->FirstChildElement("name"); el; el = el->NextSiblingElement("name"))
-      {
-        name_ = el->GetText();
-        break;
-      }
-      // Get license texts, where there may be multiple elements for.
-      std::string tagname_license = "license";
-      for(TiXmlElement* el = root->FirstChildElement(tagname_license); el; el = el->NextSiblingElement(tagname_license ))
-      {
-        licenses_.push_back(el->GetText());
-      }
-      // check if package is a metapackage
-      for(TiXmlElement* el = root->FirstChildElement("export"); el; el = el->NextSiblingElement("export"))
-      {
-        if(el->FirstChildElement("metapackage"))
-        {
-          is_metapackage_ = true;
-          break;
-        }
-      }
+      is_metapackage_ = true;
+      break;
     }
+  }
+}
 
-    bool isStack() const
-    {
-      return manifest_name_ == MANIFEST_TAG_STACK || (is_wet_package_ && is_metapackage_);
-    }
+bool Stackage::isStack() const
+{
+  return manifest_name_ == MANIFEST_TAG_STACK || (is_wet_package_ && is_metapackage_);
+}
 
-    bool isPackage() const
-    {
-      return manifest_name_ == MANIFEST_TAG_PACKAGE || (is_wet_package_ && !is_metapackage_);
-    }
-
-};
+bool Stackage::isPackage() const
+{
+  return manifest_name_ == MANIFEST_TAG_PACKAGE || (is_wet_package_ && !is_metapackage_);
+}
 
 class DirectoryCrawlRecord
 {
